+++ src/atom/atom.rs
@@ -17,17 +17,21 @@ use super::elst::Elst;
 use super::fdel::Fdel;
 use super::fecr::Fecr;
 use super::fiin::Fiin;
+use super::fire::Fire;
 use super::fpar::Fpar;
 use super::freespace::{Free, Skip};
 use super::frma::Frma;
 use super::ftyp::Ftyp;
+use super::gitn::Gitn;
 use super::hdlr::Hdlr;
 use super::header::Header;
 use super::hmhd::Hmhd;
 use super::ignore::Ignore;
 use super::iinf::Iinf;
 use super::iloc::Iloc;
+use super::imif::Imif;
 use super::infe::Infe;
+use super::ipmc::Ipmc;
 use super::ipro::Ipro;
 use super::kind::Kind;
 use super::mdat::Mdat;
@@ -86,6 +90,7 @@ use super::xml::Xml;
 /// An atom is a container for data in an MP4 file.
 #[derive(Debug, Clone)]
 pub enum Atom {
+    // Container
     Bxml(Bxml), // Meta
     Co64(Co64), // Stbl
     Cprt(Cprt), // Udta
@@ -98,16 +103,20 @@ pub enum Atom {
     Fdel(Fdel), // Iinf
     Fecr(Fecr), // Fiin
     Fiin(Fiin), // Iinf
+    Fire(Fire), // Meta
     Fpar(Fpar), // Meta
     Free(Free),
     Frma(Frma), // Sinf, Rinf, Cinf
     Ftyp(Ftyp),
+    Gitn(Gitn), // Meta
     Hdlr(Hdlr), // Moov
     Hmhd(Hmhd), // Moov
     Ignore(Ignore),
     Iinf(Iinf), // Meta
     Iloc(Iloc), // Meta
+    Imif(Imif), // Meta
     Infe(Infe), // Meta
+    Ipmc(Ipmc), // Meta, Moov
     Ipro(Ipro), // Meta
     Mdat(Mdat),
     Mdhd(Mdhd), // Moov
@@ -203,209 +212,98 @@ impl Atom {
     pub fn parse(f: &mut Mp4File) -> Result<Self, &'static str> {
         let_ok!(header, Header::parse(f), "Unable to read Atom header.");
 
+        // Create the error message for the match statement.
+        let p_err = format!("Unable to parse Kind::{}", header.kind);
+        let pe = p_err.as_str();
+
         match header.kind {
-            Kind::Bxml => Ok(Self::Bxml(
-                Bxml::parse(f, header).expect("Unable to parse Kind::Bxml"),
-            )),
-            Kind::Co64 => Ok(Self::Co64(
-                Co64::parse(f, header).expect("Unable to parse Kind::Co64"),
-            )),
-            Kind::Cslg => Ok(Self::Cslg(
-                Cslg::parse(f, header).expect("Unable to parse Kind::Cslg"),
-            )),
-            Kind::Cprt => Ok(Self::Cprt(
-                Cprt::parse(f, header).expect("Unable to parse Kind::Cprt"),
-            )),
-            Kind::Ctts => Ok(Self::Ctts(
-                Ctts::parse(f, header).expect("Unable to parse Kind::Ctts"),
-            )),
-            Kind::Dinf => Ok(Self::Dinf(Dinf::parse(f, header))),
-            Kind::Dref => Ok(Self::Dref(
-                Dref::parse(f, header).expect("Unable to parse Kind::Dref"),
-            )),
-            Kind::Edts => Ok(Self::Edts(Edts::parse(f, header))),
-            Kind::Elst => Ok(Self::Elst(
-                Elst::parse(f, header).expect("Unable to parse Kind::Elst"),
-            )),
-            Kind::Fdel => Ok(Self::Fdel(
-                Fdel::parse(f, header).expect("Unable to parse Kind::Fdel"),
-            )),
-            Kind::Fecr => Ok(Self::Fecr(
-                Fecr::parse(f, header).expect("Unable to parse Kind::Fecr"),
-            )),
-            Kind::Fiin => Ok(Self::Fiin(
-                Fiin::parse(f, header).expect("Unable to parse Kind::Fiin"),
-            )),
-            Kind::Fpar => Ok(Self::Fpar(
-                Fpar::parse(f, header).expect("Unable to parse Kind::Fpar"),
-            )),
-            Kind::Free => Ok(Self::Free(
-                Free::parse(f, header).expect("Unable to parse Kind::Free"),
-            )),
-            Kind::Frma => Ok(Self::Frma(
-                Frma::parse(f, header).expect("Unable to parse Kind::Frma"),
-            )),
-            Kind::Ftyp => Ok(Self::Ftyp(
-                Ftyp::parse(f, header).expect("Unable to parse Kind::Ftyp"),
-            )),
-            Kind::Hdlr => Ok(Self::Hdlr(
-                Hdlr::parse(f, header).expect("Unable to parse Kind::Hdlr"),
-            )),
-            Kind::Hmhd => Ok(Self::Hmhd(
-                Hmhd::parse(f, header).expect("Unable to parse Kind::Hmhd"),
-            )),
-            Kind::Iinf => Ok(Self::Iinf(
-                Iinf::parse(f, header).expect("Unable to parse Kind::Iinf"),
-            )),
-            Kind::Iloc => Ok(Self::Iloc(
-                Iloc::parse(f, header).expect("Unable to parse Kind::Iloc"),
-            )),
-            // Kind::Imif => ,
-            // Kind::Ipmc => ,
-            Kind::Ipro => Ok(Self::Ipro(
-                Ipro::parse(f, header).expect("Unable to parse Kind::Ipro"),
-            )),
-            // Kind::Itn  => ,
-            Kind::Mdat => Ok(Self::Mdat(
-                Mdat::parse(f, header).expect("Unable to parse Kind::Mdat"),
-            )),
-            Kind::Mdhd => Ok(Self::Mdhd(
-                Mdhd::parse(f, header).expect("Unable to parse Kind::Mdhd"),
-            )),
-            Kind::Mdia => Ok(Self::Mdia(Mdia::parse(f, header))),
-            Kind::Meco => Ok(Self::Meco(Meco::parse(f, header))),
-            Kind::Mehd => Ok(Self::Mehd(
-                Mehd::parse(f, header).expect("Unable to parse Kind::Mehd"),
-            )),
-            Kind::Mere => Ok(Self::Mere(
-                Mere::parse(f, header).expect("Unable to parse Kind::Mere"),
-            )),
-            Kind::Meta => Ok(Self::Meta(
-                Meta::parse(f, header).expect("Unable to parse Kind::Meta"),
-            )),
-            Kind::Mfhd => Ok(Self::Mfhd(
-                Mfhd::parse(f, header).expect("Unable to parse Kind::Mfhd"),
-            )),
-            Kind::Mfra => Ok(Self::Mfra(Mfra::parse(f, header))),
-            Kind::Mfro => Ok(Self::Mfro(
-                Mfro::parse(f, header).expect("Unable to parse Kind::Mfro"),
-            )),
-            Kind::Minf => Ok(Self::Minf(Minf::parse(f, header))),
-            Kind::Moof => Ok(Self::Moof(Moof::parse(f, header))),
-            Kind::Moov => Ok(Self::Moov(Moov::parse(f, header))),
-            Kind::Mvex => Ok(Self::Mvex(Mvex::parse(f, header))),
-            Kind::Mvhd => Ok(Self::Mvhd(
-                Mvhd::parse(f, header).expect("Unable to parse Kind::Mvhd"),
-            )),
-            Kind::Mmhd => Ok(Self::Mmhd(Nmhd::parse(f, header))),
-            Kind::Padb => Ok(Self::Padb(
-                Padb::parse(f, header).expect("Unable to parse Kind::Padb"),
-            )),
+            Kind::Bxml => Ok(Self::Bxml(Bxml::parse(f, header).expect(pe))),
+            Kind::Co64 => Ok(Self::Co64(Co64::parse(f, header).expect(pe))),
+            Kind::Cslg => Ok(Self::Cslg(Cslg::parse(f, header).expect(pe))),
+            Kind::Cprt => Ok(Self::Cprt(Cprt::parse(f, header).expect(pe))),
+            Kind::Ctts => Ok(Self::Ctts(Ctts::parse(f, header).expect(pe))),
+            Kind::Dinf => Ok(Self::Dinf(Dinf::parse(f, header).expect(pe))),
+            Kind::Dref => Ok(Self::Dref(Dref::parse(f, header).expect(pe))),
+            Kind::Edts => Ok(Self::Edts(Edts::parse(f, header).expect(pe))),
+            Kind::Elst => Ok(Self::Elst(Elst::parse(f, header).expect(pe))),
+            Kind::Fdel => Ok(Self::Fdel(Fdel::parse(f, header).expect(pe))),
+            Kind::Fecr => Ok(Self::Fecr(Fecr::parse(f, header).expect(pe))),
+            Kind::Fiin => Ok(Self::Fiin(Fiin::parse(f, header).expect(pe))),
+            Kind::Fire => Ok(Self::Fire(Fire::parse(f, header).expect(pe))),
+            Kind::Fpar => Ok(Self::Fpar(Fpar::parse(f, header).expect(pe))),
+            Kind::Free => Ok(Self::Free(Free::parse(f, header).expect(pe))),
+            Kind::Frma => Ok(Self::Frma(Frma::parse(f, header).expect(pe))),
+            Kind::Ftyp => Ok(Self::Ftyp(Ftyp::parse(f, header).expect(pe))),
+            Kind::Gitn => Ok(Self::Gitn(Gitn::parse(f, header).expect(pe))),
+            Kind::Hdlr => Ok(Self::Hdlr(Hdlr::parse(f, header).expect(pe))),
+            Kind::Hmhd => Ok(Self::Hmhd(Hmhd::parse(f, header).expect(pe))),
+            Kind::Iinf => Ok(Self::Iinf(Iinf::parse(f, header).expect(pe))),
+            Kind::Iloc => Ok(Self::Iloc(Iloc::parse(f, header).expect(pe))),
+            Kind::Imif => Ok(Self::Imif(Imif::parse(f, header).expect(pe))),
+            Kind::Ipmc => Ok(Self::Ipmc(Ipmc::parse(f, header).expect(pe))),
+            Kind::Ipro => Ok(Self::Ipro(Ipro::parse(f, header).expect(pe))),
+            Kind::Mdat => Ok(Self::Mdat(Mdat::parse(f, header).expect(pe))),
+            Kind::Mdhd => Ok(Self::Mdhd(Mdhd::parse(f, header).expect(pe))),
+            Kind::Mdia => Ok(Self::Mdia(Mdia::parse(f, header).expect(pe))),
+            Kind::Meco => Ok(Self::Meco(Meco::parse(f, header).expect(pe))),
+            Kind::Mehd => Ok(Self::Mehd(Mehd::parse(f, header).expect(pe))),
+            Kind::Mere => Ok(Self::Mere(Mere::parse(f, header).expect(pe))),
+            Kind::Meta => Ok(Self::Meta(Meta::parse(f, header).expect(pe))),
+            Kind::Mfhd => Ok(Self::Mfhd(Mfhd::parse(f, header).expect(pe))),
+            Kind::Mfra => Ok(Self::Mfra(Mfra::parse(f, header).expect(pe))),
+            Kind::Mfro => Ok(Self::Mfro(Mfro::parse(f, header).expect(pe))),
+            Kind::Minf => Ok(Self::Minf(Minf::parse(f, header).expect(pe))),
+            Kind::Moof => Ok(Self::Moof(Moof::parse(f, header).expect(pe))),
+            Kind::Moov => Ok(Self::Moov(Moov::parse(f, header).expect(pe))),
+            Kind::Mvex => Ok(Self::Mvex(Mvex::parse(f, header).expect(pe))),
+            Kind::Mvhd => Ok(Self::Mvhd(Mvhd::parse(f, header).expect(pe))),
+            Kind::Mmhd => Ok(Self::Mmhd(Nmhd::parse(f, header).expect(pe))),
+            Kind::Padb => Ok(Self::Padb(Padb::parse(f, header).expect(pe))),
             // Kind::Paen => ,
-            Kind::Pdin => Ok(Self::Pdin(
-                Pdin::parse(f, header).expect("Unable to parse Kind::Pdin"),
-            )),
-            Kind::Pitm => Ok(Self::Pitm(
-                Pitm::parse(f, header).expect("Unable to parse Kind::Pitm"),
-            )),
-            Kind::Sbgp => Ok(Self::Sbgp(
-                Sbgp::parse(f, header).expect("Unable to parse Kind::Sbgp"),
-            )),
-            Kind::Schi => Ok(Self::Schi(Schi::parse(f, header))),
-            Kind::Schm => Ok(Self::Schm(
-                Schm::parse(f, header).expect("Unable to parse Kind::Schm"),
-            )),
-            Kind::Sdtp => Ok(Self::Sdtp(
-                Sdtp::parse(f, header).expect("Unable to parse Kind::Sdtp"),
-            )),
+            Kind::Pdin => Ok(Self::Pdin(Pdin::parse(f, header).expect(pe))),
+            Kind::Pitm => Ok(Self::Pitm(Pitm::parse(f, header).expect(pe))),
+            Kind::Sbgp => Ok(Self::Sbgp(Sbgp::parse(f, header).expect(pe))),
+            Kind::Schi => Ok(Self::Schi(Schi::parse(f, header).expect(pe))),
+            Kind::Schm => Ok(Self::Schm(Schm::parse(f, header).expect(pe))),
+            Kind::Sdtp => Ok(Self::Sdtp(Sdtp::parse(f, header).expect(pe))),
             // Kind::Sgpd => ,
-            Kind::Sinf => Ok(Self::Sinf(Sinf::parse(f, header))),
-            Kind::Skip => Ok(Self::Skip(
-                Skip::parse(f, header).expect("Unable to parse Kind::Skip"),
-            )),
-            Kind::Smhd => Ok(Self::Smhd(
-                Smhd::parse(f, header).expect("Unable to parse Kind::Smhd"),
-            )),
-            Kind::Stbl => Ok(Self::Stbl(Stbl::parse(f, header))),
-            Kind::Stco => Ok(Self::Stco(
-                Stco::parse(f, header).expect("Unable to parse Kind::Stco"),
-            )),
-            Kind::Stdp => Ok(Self::Stdp(
-                Stdp::parse(f, header).expect("Unable to parse Kind::Stdp"),
-            )),
-            Kind::Stsc => Ok(Self::Stsc(
-                Stsc::parse(f, header).expect("Unable to parse Kind::Stsc"),
-            )),
-            Kind::Stsd => Ok(Self::Stsd(
-                Stsd::parse(f, header).expect("Unable to parse Kind::Stsd"),
-            )),
-            Kind::Stsh => Ok(Self::Stsh(
-                Stsh::parse(f, header).expect("Unable to parse Kind::Stsh"),
-            )),
-            Kind::Stss => Ok(Self::Stss(
-                Stss::parse(f, header).expect("Unable to parse Kind::Stss"),
-            )),
-            Kind::Stsz => Ok(Self::Stsz(
-                Stsz::parse(f, header).expect("Unable to parse Kind::Stsz"),
-            )),
-            Kind::Stts => Ok(Self::Stts(
-                Stts::parse(f, header).expect("Unable to parse Kind::Stts"),
-            )),
-            Kind::Stz2 => Ok(Self::Stz2(
-                Stz2::parse(f, header).expect("Unable to parse Kind::Stz2"),
-            )),
+            Kind::Sinf => Ok(Self::Sinf(Sinf::parse(f, header).expect(pe))),
+            Kind::Skip => Ok(Self::Skip(Skip::parse(f, header).expect(pe))),
+            Kind::Smhd => Ok(Self::Smhd(Smhd::parse(f, header).expect(pe))),
+            Kind::Stbl => Ok(Self::Stbl(Stbl::parse(f, header).expect(pe))),
+            Kind::Stco => Ok(Self::Stco(Stco::parse(f, header).expect(pe))),
+            Kind::Stdp => Ok(Self::Stdp(Stdp::parse(f, header).expect(pe))),
+            Kind::Stsc => Ok(Self::Stsc(Stsc::parse(f, header).expect(pe))),
+            Kind::Stsd => Ok(Self::Stsd(Stsd::parse(f, header).expect(pe))),
+            Kind::Stsh => Ok(Self::Stsh(Stsh::parse(f, header).expect(pe))),
+            Kind::Stss => Ok(Self::Stss(Stss::parse(f, header).expect(pe))),
+            Kind::Stsz => Ok(Self::Stsz(Stsz::parse(f, header).expect(pe))),
+            Kind::Stts => Ok(Self::Stts(Stts::parse(f, header).expect(pe))),
+            Kind::Stz2 => Ok(Self::Stz2(Stz2::parse(f, header).expect(pe))),
             // Kind::Subs => ,
-            Kind::Tfhd => Ok(Self::Tfhd(
-                Tfhd::parse(f, header).expect("Unable to parse Kind::Tfhd"),
-            )),
-            Kind::Tfra => Ok(Self::Tfra(
-                Tfra::parse(f, header).expect("Unable to parse Kind::Tfra"),
-            )),
-            Kind::Tkhd => Ok(Self::Tkhd(
-                Tkhd::parse(f, header).expect("Unable to parse Kind::Tkhd"),
-            )),
-            Kind::Traf => Ok(Self::Traf(Traf::parse(f, header))),
-            Kind::Trak => Ok(Self::Trak(Trak::parse(f, header))),
-            Kind::Tref => Ok(Self::Tref(
-                Tref::parse(f, header).expect("Unable to parse Kind::Tref"),
-            )),
-            Kind::Trgr => Ok(Self::Trgr(
-                Trgr::parse(f, header).expect("Unable to parse Kind::Trgr"),
-            )),
-            Kind::Trex => Ok(Self::Trex(
-                Trex::parse(f, header).expect("Unable to parse Kind::Trex"),
-            )),
-            Kind::Trun => Ok(Self::Trun(
-                Trun::parse(f, header).expect("Unable to parse Kind::Trun"),
-            )),
-            Kind::Tsel => Ok(Self::Tsel(
-                Tsel::parse(f, header).expect("Unable to parse Kind::Tsel"),
-            )),
-            Kind::Udta => Ok(Self::Udta(Udta::parse(f, header))),
-            Kind::Uuid => Ok(Self::Uuid(
-                Uuid::parse(f, header).expect("Unable to parse Kind::Uuid"),
-            )),
-            Kind::Vmhd => Ok(Self::Vmhd(
-                Vmhd::parse(f, header).expect("Unable to parse Kind::Vmhd"),
-            )),
-            Kind::Xml => Ok(Self::Xml(
-                Xml::parse(f, header).expect("Unable to parse Kind::Xml"),
-            )),
+            Kind::Tfhd => Ok(Self::Tfhd(Tfhd::parse(f, header).expect(pe))),
+            Kind::Tfra => Ok(Self::Tfra(Tfra::parse(f, header).expect(pe))),
+            Kind::Tkhd => Ok(Self::Tkhd(Tkhd::parse(f, header).expect(pe))),
+            Kind::Traf => Ok(Self::Traf(Traf::parse(f, header).expect(pe))),
+            Kind::Trak => Ok(Self::Trak(Trak::parse(f, header).expect(pe))),
+            Kind::Tref => Ok(Self::Tref(Tref::parse(f, header).expect(pe))),
+            Kind::Trgr => Ok(Self::Trgr(Trgr::parse(f, header).expect(pe))),
+            Kind::Trex => Ok(Self::Trex(Trex::parse(f, header).expect(pe))),
+            Kind::Trun => Ok(Self::Trun(Trun::parse(f, header).expect(pe))),
+            Kind::Tsel => Ok(Self::Tsel(Tsel::parse(f, header).expect(pe))),
+            Kind::Udta => Ok(Self::Udta(Udta::parse(f, header).expect(pe))),
+            Kind::Uuid => Ok(Self::Uuid(Uuid::parse(f, header).expect(pe))),
+            Kind::Vmhd => Ok(Self::Vmhd(Vmhd::parse(f, header).expect(pe))),
+            Kind::Xml => Ok(Self::Xml(Xml::parse(f, header).expect(pe))),
             // Kind::Strk => ,
             // Kind::Stri => ,
             // Kind::Strd => ,
-            Kind::Url => Ok(Self::Url(
-                Url::parse(f, header).expect("Unable to parse Kind::Url"),
-            )),
-            Kind::Urn => Ok(Self::Urn(
-                Urn::parse(f, header).expect("Unable to parse Kind::Urn"),
-            )),
+            Kind::Url => Ok(Self::Url(Url::parse(f, header).expect(pe))),
+            Kind::Urn => Ok(Self::Urn(Urn::parse(f, header).expect(pe))),
             Kind::Unrecognized(_) => Ok(Self::Unrecognized(
-                Unrecognized::parse(f, header).expect("Unable to parse Kind::Unrecognized"),
+                Unrecognized::parse(f, header).expect(pe),
             )),
-            _ => Ok(Self::Ignore(
-                Ignore::parse(f, header).expect("Unable to parse Kind::Ignore"),
-            )),
+            _ => Ok(Self::Ignore(Ignore::parse(f, header).expect(pe))),
         }
     }
 
@@ -427,23 +325,16 @@ impl Atom {
     /// # Panics
     ///
     /// None.
-    pub fn parse_children(f: &mut Mp4File) -> Vec<Self> {
+    pub fn parse_children(f: &mut Mp4File) -> Result<Vec<Self>, &'static str> {
         let mut atoms: Vec<Self> = Vec::new();
         loop {
             if f.offset() == f.file_size() {
                 break;
             }
 
-            match Self::parse(f) {
-                Ok(atom) => {
-                    atoms.push(atom);
-                }
-                Err(e) => {
-                    log::trace!("Atom::parse_children - parse error: ({e:?}), file: {f:?}");
-                    break;
-                }
-            }
+            let_ok!(atom, Self::parse(f), "Unable to parse atom.");
+            atoms.push(atom);
         }
-        atoms
+        Ok(atoms)
     }
 }