+++ src/atom/nmhd.rs
@@ -2,7 +2,7 @@
 
 use crate::atom::header::Header;
 use crate::mp4file::Mp4File;
-use crate::retref;
+use crate::{generic_parse, retref};
 
 #[derive(Debug, Clone)]
 pub struct Nmhd {
@@ -10,12 +10,6 @@ pub struct Nmhd {
 }
 
 impl Nmhd {
-    pub fn parse(f: &mut Mp4File, mut header: Header) -> Self {
-        header.parse_version(f);
-        header.parse_flags(f);
-
-        Self { header }
-    }
-
+    generic_parse!(Nmhd);
     retref!(header, Header);
 }